Update Elixr and Erlang and fix breaking APIs
This updates Elixr and Erlang to be the most recent locally.

I bumped the minimum Elixir version to `1.11`. The reason for this is
because from my experience with open source Elixir only supporting 3 or
4 versions back is usually good enough.

I had to update the crypto hmac function calls as OTP 24 removed
functions.

Timex had to be updated to get all the test passing. However, all the
deps are out of data, which I can update in a follow up commit.